repo_cloner/requirements.txt
fastapi
uvicorn
gitpython
pydantic